+++ src/Svg.php
@@ -61,7 +61,7 @@ class Svg
     protected function parseShapes(): void
     {
         $matches = [];
-        preg_match_all('/<(path|rect|circle|ellipse|polygon)[^\/]+\/>/', $this->contents, $matches);
+        preg_match_all('/<(path|rect|circle|ellipse|polygon)[^\/]+\/>/sm', $this->contents, $matches);
 
         foreach ($matches[0] ?? [] as $match) {
             $type = preg_replace('/<([^\s]+).*/', '$1', $match);
@@ -83,7 +83,6 @@ class Svg
                     break;
                 default:
                     throw new UnsupportedElementException();
-                    break;
             }
             $this->shapes[] = $shape;
         }